+++ types/sensative/presence/uplink.spec.js
@@ -1,0 +1,79 @@
+const chai = require("chai");
+const { validate } = require("jsonschema");
+const rewire = require("rewire");
+const utils = require("test-utils");
+
+const { assert } = chai;
+
+describe("Sensative presence Uplink", () => {
+  let defaultSchema = null;
+  let lifecycleSchema = null;
+  let alarmSchema = null;
+  let consume = null;
+  before((done) => {
+    const script = rewire("./uplink.js");
+    consume = utils.init(script);
+    utils
+      .loadSchema(`${__dirname}/default.schema.json`)
+      .then((parsedSchema) => {
+        defaultSchema = parsedSchema;
+      });
+    utils
+      .loadSchema(`${__dirname}/lifecycle.schema.json`)
+      .then((parsedSchema) => {
+        lifecycleSchema = parsedSchema;
+        done();
+      });
+    utils.loadSchema(`${__dirname}/alarm.schema.json`).then((parsedSchema) => {
+      alarmSchema = parsedSchema;
+    });
+  });
+
+  describe("consume()", () => {
+    it("should decode the sensative presence payload", () => {
+      const data = {
+        data: {
+          port: 1,
+          payloadHex: "ffff01601100001500",
+        },
+      };
+
+      utils.expectEmits((type, value) => {
+        assert.equal(type, "sample");
+        assert.isNotNull(value);
+        assert.typeOf(value.data, "object");
+
+        assert.equal(value.topic, "lifecycle");
+        assert.equal(value.data.batteryLevel, 96);
+        assert.equal(value.data.historySeqNr, 65535);
+        assert.equal(value.data.prevHistSeqNr, 65535);
+
+        validate(value.data, lifecycleSchema, { throwError: true });
+      });
+
+      utils.expectEmits((type, value) => {
+        assert.equal(type, "sample");
+        assert.isNotNull(value);
+        assert.typeOf(value.data, "object");
+
+        assert.equal(value.topic, "alarm");
+        assert.equal(value.data.closeProximityAlarm, false);
+
+        validate(value.data, alarmSchema, { throwError: true });
+      });
+
+      utils.expectEmits((type, value) => {
+        assert.equal(type, "sample");
+        assert.isNotNull(value);
+        assert.typeOf(value.data, "object");
+
+        assert.equal(value.topic, "default");
+        assert.equal(value.data.doorCount, 0);
+
+        validate(value.data, defaultSchema, { throwError: true });
+      });
+
+      consume(data);
+    });
+  });
+});